refactor(Option): 메뉴 옵션별 분기 확인 메서드 제거 및 dto 도입
